Convert `ocrs_models.dataset` module into a package
This module was getting large. Convert it into a package with separate modules
for individual datasets, the CLI interface and shared utilities.
