+++ ocrs_models/eval_detection.py
@@ -8,7 +8,7 @@ import torch
 from torchvision.transforms.functional import InterpolationMode, resize, to_pil_image
 
 from .models import DetectionModel
-from .datasets import SHRINK_DISTANCE
+from .datasets.util import SHRINK_DISTANCE
 from .postprocess import draw_quads, expand_quads, extract_cc_quads
 from .train_detection import mask_size
